Remove 'flat' directory, move everything to root.
