+++ eslint.config.js
@@ -1,6 +1,6 @@
-import ignores from "./flat/ignores.js";
+import ignores from "./ignores.js";
 import meta from "@cto.af/eslint-plugin-meta";
-import module from "./flat/module.js";
+import module from "./module.js";
 import stylistic from "@stylistic/eslint-plugin";
 import { plugin as ts } from "typescript-eslint";
