+++ globals.js
