+++ test/flat-module/eslint.config.js
@@ -1,7 +1,7 @@
-import ignores from "@peggyjs/eslint-config/flat/ignores.js";
-import mocha from "@peggyjs/eslint-config/flat/mocha.js";
-import module from "@peggyjs/eslint-config/flat/module.js";
-import ts from "@peggyjs/eslint-config/flat/ts.js";
+import ignores from "@peggyjs/eslint-config/ignores.js";
+import mocha from "@peggyjs/eslint-config/mocha.js";
+import module from "@peggyjs/eslint-config/module.js";
+import ts from "@peggyjs/eslint-config/ts.js";
 
 export default [
   ...ignores,
