+++ test/flat/eslint.config.mjs
@@ -1,6 +1,6 @@
-import commonjs from "@peggyjs/eslint-config/flat/commonjs.js";
-import mocha from "@peggyjs/eslint-config/flat/mocha.js";
-import ts from "@peggyjs/eslint-config/flat/ts.js";
+import commonjs from "@peggyjs/eslint-config/commonjs.js";
+import mocha from "@peggyjs/eslint-config/mocha.js";
+import ts from "@peggyjs/eslint-config/ts.js";
 
 export default [
   ...commonjs,
